batch-wise processing of files to conserve RAM
