--- ddh_tactus/ddh_proc.py
@@ -1,107 +1,0 @@
-#!/usr/bin/env python
-# coding: utf-8
-
-import epygram as epg
-import xarray as xr
-import numpy as np
-import pandas as pd
-import glob
-import dask.bag as db
-from dask.distributed import Client, progress
-
-
-# -- initialize environemnt
-epg.init_env()
-
-# -- Files list
-dir1 = '/ec/res4/scratch/swe7088/deode/ddh_mat_CY49t2_HARMONIE_AROME_LES_input_Paris_200m_linear_20230820/archive/2023/08/20/12/mbr000/'
-file_list = glob.glob(dir1+'DHFDLDEOD+*s')
-file_list.sort()
-file_list = file_list
-file1 = file_list[0]
-
-
-# -- Setup up metada from first file
-with epg.formats.resource(file_list[0], openmode='r', fmt='DDHLFA') as res:
-    geom = res.domains['geometry']
-    n_levels = len(res.readfield('VPP0')[0].geometry.vcoordinate.levels)
-
-    # Pre-calculate mapping indices
-    jgl = np.array([d['jgl'] for d in geom])
-    jlon = np.array([d['jlon'] for d in geom])
-    lats = np.array([d['lat'].get() for d in geom])
-    lons = np.array([d['lon'].get() for d in geom])
-
-# -- grid data
-jgl_un = np.unique(jgl)
-jlon_un = np.unique(jlon)
-lats = np.array(lats)
-lons = np.array(lons)
-
-
-lons_grid = lons.reshape((jlon_un.size, jgl_un.size), order='F')
-lats_grid = lats.reshape((jlon_un.size, jgl_un.size), order='F')
-
-articles = pd.read_table('ddh_article_list3').values.flatten()
-
-# -- mapping dictionary for indices
-R = jlon_un.size
-map_ij = {}
-for i in range(len(jgl)):
-    map_ij[i]=(i%R, int(i/R))
-
-def proc_one_file(file_path):
-    with epg.formats.resource(file_path, openmode='r', fmt='DDHLFA') as res_loc:
-        time_now = res_loc.validity.get()
-        file_data = {}
-        for i, article in enumerate(articles):
-            field_loc = res_loc.readfield(article)
-            data_loc = np.zeros([n_levels, jlon_un.size, jgl_un.size])
-
-            for j, domain in enumerate(field_loc):
-                # geom = domain.geometry.grid
-                ix, jx = map_ij[j]
-                data_loc[:, ix, jx] = domain.data
-
-            file_data[article] = (['level', 'jlon', 'jgl'], data_loc)
-
-    # -- Make data set and return
-    ds = xr.Dataset(
-        data_vars=file_data,
-        coords={
-            'time': [time_now],
-            'level': np.arange(n_levels) + 1,
-            'jlon': jlon_un,
-            'jgl': jgl_un,
-            'lat': (['jlon', 'jgl'], lats_grid),
-            'lon': (['jlon', 'jgl'], lons_grid)
-        }
-    )
-    return ds
-
-# Use Dask Bag to parallelize
-if __name__ == "__main__":
-    # Start a local cluster (adjust n_workers based on your RAM/CPUs)
-    client = Client(n_workers=32, threads_per_worker=1)
-
-    print(f"Starting parallel processing of {len(file_list)} files...")
-
-    # Create bag, map the function, and compute
-    # npartitions should be balanced; ~100 files per partition is usually healthy
-    b = db.from_sequence(file_list, npartitions=12)
-    datasets = b.map(proc_one_file).compute()
-
-    # 5. Concatenate all datasets along the time dimension
-    # Using 'combine_by_coords' or 'concat'
-    ds = xr.concat(datasets, dim='time').sortby('time')
-
-    print("Processing complete.")
-
-    # Sanity check
-    print(ds)
-
-    print("Saving to zarr")
-    ds.to_zarr('DDH_out.zarr', 'w')
-
-
-
